backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pet-api"
version = "0.1.0"
description = "Pet API - FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "fastapi",
    "uvicorn[standard]>=0.27.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "httpx>=0.26.0",
    "google-genai",
    "pillow>=10.0.0",
    "qrcode[pil]>=7.4.0",
    "google>=3.0.0",
    "python-multipart>=0.0.22",
    "boto3>=1.34.0",
    "alembic>=1.13.0",
    "transformers>=4.45.0",
    "torch>=2.0.0",
    "accelerate>=0.33.0",
    "streamlink>=8.1.2",
    "opencv-python-headless>=4.8.0",
    "python-jose[cryptography]>=3.5.0",
    "bcrypt>=5.0.0",
    "moviepy>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app", "tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
